+++ MapboxNavigation/NextBannerView.swift
@@ -1,5 +1,6 @@
 import UIKit
 import MapboxDirections
+import MapboxCoreNavigation
 
 /// :nodoc:
 @objc(MBNextInstructionLabel)
@@ -52,7 +53,7 @@ open class NextBannerView: UIView {
     override open func prepareForInterfaceBuilder() {
         super.prepareForInterfaceBuilder()
         maneuverView.isEnd = true
-        instructionLabel.instruction = [VisualInstructionComponent(type: .destination, text: "Next step", imageURL: nil)]
+        instructionLabel.instruction = [VisualInstructionComponent(type: .destination, text: "Next step", imageURL: nil, maneuverType: .none, maneuverDirection: .none)]
     }
     
     func setupLayout() {
@@ -69,6 +70,49 @@ open class NextBannerView: UIView {
         instructionLabel.leftAnchor.constraint(equalTo: leftAnchor, constant: 70).isActive = true
         instructionLabel.centerYAnchor.constraint(equalTo: centerYAnchor).isActive = true
         instructionLabel.rightAnchor.constraint(equalTo: rightAnchor, constant: -16).isActive = true
+    }
+    
+    func shouldShowNextBanner(for routeProgress: RouteProgress) -> Bool {
+        guard let upcomingStep = routeProgress.currentLegProgress.upComingStep, let nextStep = routeProgress.currentLegProgress.stepAfter(upcomingStep) else {
+            return false
+        }
+        
+        // If the followon step is short and the user is near the end of the current step, show the nextBanner.
+        guard nextStep.expectedTravelTime <= RouteControllerHighAlertInterval * RouteControllerLinkedInstructionBufferMultiplier,
+            upcomingStep.expectedTravelTime <= RouteControllerHighAlertInterval * RouteControllerLinkedInstructionBufferMultiplier else {
+                return false
+        }
+        
+        guard let _ = upcomingStep.instructionsDisplayedAlongStep?.last else { return false }
+        
+        return true
+    }
+    
+    public func update(for routeProgress: RouteProgress) {
+        guard shouldShowNextBanner(for: routeProgress) else {
+            hide()
+            return
+        }
+        
+        guard let instruction = routeProgress.currentLegProgress.upComingStep?.instructionsDisplayedAlongStep?.last else { return }
+        
+        maneuverView.visualInstruction = instruction
+        instructionLabel.instruction = instruction.primaryTextComponents
+        show()
+    }
+    
+    public func show() {
+        guard isHidden else { return }
+        UIView.defaultAnimation(0.3, animations: {
+            self.isHidden = false
+        }, completion: nil)
+    }
+    
+    public func hide() {
+        guard !isHidden else { return }
+        UIView.defaultAnimation(0.3, animations: {
+            self.isHidden = true
+        }, completion: nil)
     }
     
 }